% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBaySVG_main.R
\name{citgtest}
\alias{citgtest}
\title{single gene process}
\usage{
citgtest(
  g,
  result,
  c_alpha,
  spelist,
  acrate,
  gamma1,
  gamma2,
  min_iter,
  max_iter,
  speci_iter,
  use_covariate = TRUE,
  a_pi,
  b_pi,
  cp,
  dp,
  cq,
  dq,
  ak_domain,
  a_phi_single,
  b_phi_single,
  eta_sigma_single,
  psi_sigma_single
)
}
\arguments{
\item{g}{:gene number}

\item{result}{: preprocessed data produced by CTIG}

\item{c_alpha}{A list of length m, each element being a covariates matrix.}

\item{spelist}{A list of length m. Each element is a sublist containing:
\itemize{
\item{expression_matrix: Matrix (spots x genes)}
\item{location_matrix: Matrix (spots x coordinates)}
}}

\item{acrate}{:the minimum threshold for EBLO convergence.}

\item{gamma1}{:hyperparameter of the slab and spike prior of beta. Default is sqrt(0.001).}

\item{gamma2}{:hyperparameter of the bi-level structure of alpha. We set gamma2 to 0.01, 0.005, and 0.001 for more than four, three, and two or fewer slices, respectively.}

\item{min_iter}{: the minimum number of iteration.}

\item{max_iter}{: the maximum number of iteration.}

\item{speci_iter}{: Number of times the early convergence criterion is met.}

\item{use_covariate}{:the logistic variable of using covariate. Default is TRUE which use covariates.}

\item{a_pi, b_pi}{Numeric. Hyperparameters for the Beta prior of the zero-inflation probability \eqn{\pi}. Defaults are 1 and 1.}

\item{cp}{:hyperparameter of alpha's beta prior}

\item{dp}{:hyperparameter of alpha's beta prior}

\item{cq}{:hyperparameter of u's beta prior}

\item{dq}{:hyperparameter of u's beta prior}

\item{ak_domain}{:Hyperparameter of varience of beta to control the slab and spike prior for different degree. Default are 0.08,0.05,0.04 and 0.03 for degree 1-4, respectively.}

\item{a_phi_single}{:Hyperparameter of the shape parameter \eqn{\phi}. Default is 0.001.}

\item{b_phi_single}{:Hyperparameter of the shape parameter \eqn{\phi}. Default is 0.001.}

\item{eta_sigma_single}{:Hyperparameter of variance of \eqn{\eta}, the baseline gene expression. Default is 1.}

\item{psi_sigma_single}{:Hyperparameter of variance of \eqn{\psi}, controlling the contribution of covariates. Default is 1.}
}
\value{
the posterior mean of all parameters for gene g.
}
\description{
single gene process
}
