% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBaySVG_main.R
\name{VIZINB}
\alias{VIZINB}
\title{implement function}
\usage{
VIZINB(
  y,
  splinevalue,
  samplenumber,
  splinelevel,
  acrate = 0.01,
  ak = 1,
  dp = 1.8,
  cp = 0.2,
  c_alpha,
  knum = 6,
  cq = 1,
  dq = 1,
  a_phi_single = 0.001,
  b_phi_single = 0.001,
  a_pi = 1,
  b_pi = 1,
  eta_sigma_single = 1,
  psi_sigma_single = 1,
  gamma1 = sqrt(0.001),
  gamma2 = 0.01,
  min_iter = 30,
  max_iter = 200,
  speci_iter = 12,
  use_covariate = TRUE
)
}
\arguments{
\item{y}{：list consist of a gene expression in m datasets}

\item{splinevalue}{:list consist of basic functions in m datasets with predetermined degree}

\item{samplenumber}{:vector consist of gene numbers in m datasets}

\item{splinelevel}{:predetermined degree}

\item{acrate}{:threshold value for determining ELBO convergence}

\item{ak}{:hyperparameter of varience of beta to control the slab and spike prior}

\item{dp}{:hyperparameter of alpha's beta prior}

\item{cp}{:hyperparameter of alpha's beta prior}

\item{c_alpha}{:list consist of coveriates in m datasets}

\item{knum}{:the dimension of coveriates}

\item{cq}{:hyperparameter of u's beta prior}

\item{dq}{:hyperparameter of u's beta prior}

\item{a_phi_single}{:Hyperparameter of the shape parameter \eqn{\phi}. Default is 0.001.}

\item{b_phi_single}{:Hyperparameter of the shape parameter \eqn{\phi}. Default is 0.001.}

\item{a_pi, b_pi}{Numeric. Hyperparameters for the Beta prior of the zero-inflation probability \eqn{\pi}. Defaults are 1 and 1.}

\item{eta_sigma_single}{:Hyperparameter of variance of \eqn{\eta}, the baseline gene expression. Default is 1.}

\item{psi_sigma_single}{:Hyperparameter of variance of \eqn{\psi}, controlling the contribution of covariates. Default is 1.}

\item{gamma1}{:hyperparameter of the slab and spike prior of beta. Default is sqrt(0.001).}

\item{gamma2}{:hyperparameter of the bi-level structure of alpha. We set gamma2 to 0.01, 0.005, and 0.001 to represent scenarios integrating more than four, three, and two or fewer slices, respectively.}

\item{min_iter}{: the minimum number of iteration.}

\item{max_iter}{: the maximum number of iteration.}

\item{speci_iter}{: Number of times the early convergence criterion is met}

\item{use_covariate}{:the logistic variable of using covariate. Default is TRUE which use covariates.}
}
\value{
the posterior mean of all the parameters
}
\description{
implement function
}
